src/Code/requirements.txt
argcomplete==1.10.0
beautifulsoup4==4.8.0
chardet==3.0.4
click==7.1.2
docx2txt==0.8
dominate==2.6.0
EbookLib==0.17.1
extract-msg==0.23.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
IMAPClient==2.1.0
indic-nlp-library==0.71
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.1
MarkupSafe==1.1.1
Morfessor==2.0.6
numpy==1.19.3
olefile==0.46
Open-Tamil==0.97
pandas==1.1.4
pdfminer.six==20181108
Pillow==8.0.1
polyglot==16.7.4
pycryptodome==3.9.8
PyPDF2==1.26.0
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.1
rippletagger==0.2
six==1.12.0
sortedcontainers==2.2.2
soupsieve==2.0.1
SpeechRecognition==3.8.1
textract==1.6.3
tzlocal==1.5.1
visitor==0.1.3
Werkzeug==1.0.1
xlrd==1.2.0
XlsxWriter==1.3.7
